refactored the tool directories
